# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel 6.5",
        "description": "This is a sample server for ...",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Login1"
                ],
                "summary": "Login",
                "description": "A sample ... to test out the API",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "List of all companies.",
                "description": "A sample ... to test out the API",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a company.",
                "description": "A sample ... to test out the API",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jessica Smith"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success a company creation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "Company name.",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "Date of update.",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Date of create.",
                                            "type": "string"
                                        },
                                        "Id": {
                                            "description": "Record ID.",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "name": "Jessica Smith",
                                        "updated_at": "2019-11-21 16:41:42",
                                        "created_at": "2019-11-21 16:41:42",
                                        "id": 7
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get company information",
                "description": "A sample ... to test out the API",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Company identifier.",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Company name.",
                                            "type": "string|null"
                                        },
                                        "address": {
                                            "description": "Address.",
                                            "type": "string"
                                        },
                                        "website": {
                                            "description": "Website.",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "Email.",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "Date of update.",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Date of create.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 1,
                                        "name": null,
                                        "address": null,
                                        "website": null,
                                        "email": null,
                                        "created_at": "2019-11-21 15:59:00",
                                        "updated_at": "2019-11-21 15:59:00"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/api/v1/companies/{company}": {
            "delete": {
                "tags": [
                    "Companies"
                ],
                "description": "deletes a single company based on the ID supplied",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID of company to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Company not found"
                    },
                    "204": {
                        "description": "company deleted"
                    },
                    "200": {
                        "description": "success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update a company.",
                "description": "A sample ... to test out the API",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "company": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jessica Smith"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success a company update.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "Company name.",
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "description": "Date of update.",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Date of create.",
                                            "type": "string"
                                        },
                                        "Id": {
                                            "description": "Record ID.",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "name": "Jessica Smith",
                                        "updated_at": "2019-11-21 16:41:42",
                                        "created_at": "2019-11-21 16:41:42",
                                        "id": 7
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            }
        }
    }
}